drop the `SingleMessage` adapter and the `Elfa`/`Messari` providers, trim tests to the option mapping
